Moved MainFormWrapperTest into gui.forms directory to match the corresponding class.  MainFormData now has PropertyChangeSupport
