+++ src/main/java/gui/MainFormData.java
@@ -1,7 +1,19 @@
 package gui;
 
+import java.beans.PropertyChangeSupport;
+
 public class MainFormData {
+    private PropertyChangeSupport propertyChangeSupport;
+
+    public MainFormData() {
+        propertyChangeSupport = new PropertyChangeSupport(this);
+    }
+
     public void updateVendingDisplayLabel(String displayedText) {
 
+    }
+
+    public PropertyChangeSupport getPropertyChangeSupport() {
+        return propertyChangeSupport;
     }
 }
